Added logo section to assets page
